Develop

+++ scripts/postgres/init.sql
@@ -4,7 +4,7 @@ CREATE TABLE transfers (
     id VARCHAR(36) PRIMARY KEY NOT NULL,
     account_origin_id VARCHAR NOT NULL,
     account_destination_id VARCHAR NOT NULL,
-    amount FLOAT NOT NULL,
+    amount BIGINT NOT NULL,
     created_at TIMESTAMP NOT NULL
 );
 
@@ -12,6 +12,6 @@ CREATE TABLE accounts (
     id VARCHAR(36) PRIMARY KEY NOT NULL,
     name VARCHAR NOT NULL,
     cpf VARCHAR UNIQUE NOT NULL,
-    balance FLOAT NOT NULL,
+    balance BIGINT NOT NULL,
     created_at TIMESTAMP NOT NULL
 );
